Move createShapeFile to general utilities and use it in the urban data input tool
